Move benchmarking scripts into the appropriate `bench` dir
