+++ bench/run-fast.py
