+++ bench/run_explode-js_cwe94.sh
@@ -2,13 +2,13 @@
 
 pgrep -f "neo4j" > /dev/null || neo4j start
 
-pushd bench/datasets
+pushd datasets
 
 runner run \
   --run-mode full \
   --timeout 300 \
-  --filter CWE-78 \
-  --output CWE-78 \
+  --filter CWE-94 \
+  --output CWE-94 \
   ./index.json
 
 popd
