+++ bench/run_explode-js_no-lazy-values.sh
@@ -2,7 +2,7 @@
 
 pgrep -f "neo4j" > /dev/null || neo4j start
 
-pushd bench/datasets
+pushd datasets
 
 runner run \
   --run-mode full \
@@ -13,4 +13,4 @@ runner run \
 
 popd
 
-python3 table_explode-js2.py bench/datasets/no-lazy-values/results.csv
+python3 table_explode-js2.py datasets/no-lazy-values/results.csv
