--- nodeMedic-bench-install.sh
@@ -1,2 +1,0 @@
-cd bench/NodeMedic && docker build --platform=linux/amd64 -t nodemedic-fine:latest .
-python3 -m pip install pandas~=2.2.3 tabulate~=0.9.0
