--- requirements.txt
@@ -1,2 +1,0 @@
-tabulate==0.9.0
-pandas==2.2.3
